+++ src/pages/tonieboxes/boxsetup/esp32/ESP32BoxFlashingPage.tsx
@@ -9,13 +9,9 @@ import BreadcrumbWrapper, {
 } from "../../../../components/common/StyledComponents";
 import { TonieboxesSubNav } from "../../../../components/tonieboxes/TonieboxesSubNav";
 import { Flashing } from "../../../../components/tonieboxes/boxsetup/esp32/flashing/Flashing";
-import { useGetSettingUseRevvoxFlasher } from "../../../../hooks/getsettings/useGetSettingUseRevvoxFlasher";
-import LoadingSpinner from "../../../../components/common/elements/LoadingSpinner";
 
 export const ESP32BoxFlashingPage: React.FC = () => {
     const { t } = useTranslation();
-
-    const useRevvoxFlasher = useGetSettingUseRevvoxFlasher();
 
     return (
         <>
@@ -32,7 +28,7 @@ export const ESP32BoxFlashingPage: React.FC = () => {
                     ]}
                 />
                 <StyledContent>
-                    {useRevvoxFlasher === null ? <LoadingSpinner /> : <Flashing useRevvoxFlasher={useRevvoxFlasher} />}
+                    <Flashing />
                 </StyledContent>
             </StyledLayout>
         </>